+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Multiselect: the query in the filter is cleared after pressing Enter
 - Multiselect: Hangs when changing values on heavy dashboards
+- Default value of textArea not set
 
 ## [1.132.0]